organize database functionality into a sub module
